Outsource tests, otherwise not correctly run
